fix: fix empty object getting returned for searchParams in prod build
